refactor(flame): extract BubbleManager from TechWorld (#320)
- Extract ~730 lines of bubble lifecycle management from TechWorld
  (2081→1392 lines, -33%) into a dedicated BubbleManager class
- BubbleManager is a plain Dart class (not a Flame Component) that owns
  proximity detection, physics repulsion, metaball/merged video rendering,
  audio threshold management, shader loading, and Dreamfinder avatar bridge
- Adds static chebyshevDistance() replacing 3 inline computations
- Adds 21 unit tests for the new class
- Extract _failureMessageFor() helper in main.dart
- Replace timing-dependent test delays with deterministic pumpEventQueue()
